feat(login): update style role selection
