+++ web/login/LoginSelection.js
@@ -1,6 +1,9 @@
 import React from "react";
+import { useHistory } from "react-router-dom";
+import Link from "@mui/material/Link";
+import Typography from "@mui/material/Typography";
 import { makeStyles } from "@mui/styles";
-import { ManagerImage, WorkerImage } from "common/utils/icons";
+import { ControllerImage, WorkerImage } from "common/utils/icons";
 import { Header } from "../common/Header";
 import { usePageTitle } from "../common/UsePageTitle";
 import classNames from "classnames";
@@ -18,24 +21,26 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-export function RoleSelection() {
-  usePageTitle("Sélection de rôle - Mobilic");
+export default function LoginSelection() {
+  usePageTitle("Connexion - Mobilic");
+  const history = useHistory();
   const classes = useStyles();
 
   return [
     <Header key={1} />,
     <h4 key={2} className="fr-pt-2w">
-      Quel est votre métier ?
+      Se connecter en tant que
     </h4>,
     <div key={3} className={classNames(classes.container, "fr-mx-2w")}>
       <div className="fr-card fr-enlarge-link fr-col-12 fr-col-md-4">
         <div className="fr-card__body">
           <div className="fr-card__content">
-            <h5 className="fr-card__title">
-              <a href="/signup/user">Travailleur mobile</a>
-            </h5>
+            <h4 className="fr-card__title">
+              <a href="/login">Entreprise ou salarié</a>
+            </h4>
             <p className="fr-card__desc">
-              Je suis travailleur mobile et je dois remplir le LIC
+              Je suis travailleur mobile ou gestionnaire d'une entreprise de
+              transport
             </p>
           </div>
         </div>
@@ -48,20 +53,34 @@ export function RoleSelection() {
       <div className="fr-card fr-enlarge-link fr-col-12 fr-col-md-4">
         <div className="fr-card__body">
           <div className="fr-card__content">
-            <h5 className="fr-card__title">
-              <a href="/signup/user">Gestionnaire</a>
-            </h5>
+            <h4 className="fr-card__title">
+              <a href="/controller-login">Contrôleur</a>
+            </h4>
             <p className="fr-card__desc">
-              Je suis gestionnaire d'une entreprise de transport
+              Je suis Agent public de l'État et je me connecte à mon espace
+              dédié
             </p>
           </div>
         </div>
         <div className="fr-card__header">
           <div className="fr-card__img">
-            <ManagerImage height={150} width={150} />
+            <ControllerImage height={150} width={150} />
           </div>
         </div>
       </div>
-    </div>
+    </div>,
+    <Typography key={4} my={2}>
+      Pas encore de compte ?{" "}
+      <Link
+        href="/signup"
+        onClick={e => {
+          e.preventDefault();
+          history.push("/signup");
+        }}
+      >
+        {" "}
+        Je m'inscris
+      </Link>
+    </Typography>
   ];
 }
